Stop sleeping in waitUntilAllThreadsStopped spin instead.
We occasionally have to wait for threads to finish
but 10ms seems too long and is a significant part of
total time of Zygote fork (up to 15-20%).

It turns out that all Java threads are already stopped
but we're still waiting for their OS counterparts to finish.
This shouldn't take much time so we can just spin here.

Change-Id: Ib63c0610a0786924060a644331dadaf97f65f0d5